requirements.txt
django
djangorestframework
django-cors-headers
pandas
matplotlib
reportlab
requests
PyQt5
